[AI] Move profiling code to separate module
- Move code to separate module
- Add usage clarification
- Remove unnecessary usages
