+++ default/python/AI/freeorion_tools/__init__.py
@@ -3,3 +3,4 @@ from .extend_freeorion_AI_interface import patch_interface
 from .timers import AITimer
 from freeorion_tools.fo_chat_handler import configure_debug_chat, process_chat_message
 from freeorion_tools.combine_ratings import combine_ratings
+from freeorion_tools._profile import profile
